Tests: Reorder t3* tests
