+++ tests/t310-basis-f.f90
@@ -1,37 +1,49 @@
 !-----------------------------------------------------------------------
-!
-! Header with common subroutine
-! 
-      include 't310-basis-f.h'
-!-----------------------------------------------------------------------
       program test
 
       include 'ceedf.h'
 
       integer ceed,err
-      integer p,q,d
-      parameter(p=6)
-      parameter(q=4)
-      parameter(d=2)
+      integer b,i,u,v
 
-      real*8 qref(d*q)
-      real*8 qweight(q)
-      real*8 interp(p*q)
-      real*8 grad(d*p*q)
+      integer dimn,p1d,q1d,length
+      integer*8 uoffset,voffset
+      parameter(dimn   = 2)
+      parameter(p1d    = 4)
+      parameter(q1d    = 4)
+      parameter(length = q1d**dimn)
 
-      integer b
+      real*8 uu(length)
+      real*8 vv(length)
 
       character arg*32
 
       call getarg(1,arg)
-
       call ceedinit(trim(arg)//char(0),ceed,err)
 
-      call buildmats(qref,qweight,interp,grad)
+      call ceedvectorcreate(ceed,length,u,err)
+      call ceedvectorcreate(ceed,length,v,err)
 
-      call ceedbasiscreateh1(ceed,ceed_triangle,1,p,q,interp,grad,qref,qweight,&
-     & b,err)
-      call ceedbasisview(b,err)
+      do i=1,length
+        uu(i)=1.0
+      enddo
+      uoffset=0
+      call ceedvectorsetarray(u,ceed_mem_host,ceed_use_pointer,uu,uoffset,err)
+
+      call ceedbasiscreatetensorh1lagrange(ceed,dimn,1,p1d,q1d,&
+     & ceed_gauss_lobatto,b,err)
+
+      call ceedbasisapply(b,1,ceed_notranspose,ceed_eval_interp,u,v,err)
+
+      call ceedvectorgetarrayread(v,ceed_mem_host,vv,voffset,err)
+      do i=1,length
+        if (abs(vv(i+voffset)-1.) > 1.D-15) then
+! LCOV_EXCL_START
+          write(*,*) 'v(',i,'=',vv(i+voffset),' not eqaul to 1.0'
+! LCOV_EXCL_STOP
+        endif
+      enddo
+      call ceedvectorrestorearrayread(v,vv,voffset,err)
 
       call ceedbasisdestroy(b,err)
       call ceeddestroy(ceed,err)
