+++ tests/t311-basis-f.f90
@@ -1,13 +1,17 @@
 !-----------------------------------------------------------------------
-! 
-! Header with common subroutine
-! 
-      include 't310-basis-f.h'
-!-----------------------------------------------------------------------
-      subroutine feval(x1,x2,val)
-      real*8 x1,x2,val
+      subroutine polyeval(x,n,p,uq)
+      real*8 x,y
+      integer n,i
+      real*8 p(1)
+      real*8 uq
 
-      val=x1*x1+x2*x2+x1*x2+1
+      y=p(n) 
+
+      do i=n-1,1,-1
+        y=y*x+p(i)
+      enddo
+
+      uq=y
 
       end
 !-----------------------------------------------------------------------
@@ -16,76 +20,82 @@
       include 'ceedf.h'
 
       integer ceed,err
-      integer input,output
-      integer p,q,d
-      parameter(p=6)
-      parameter(q=4)
-      parameter(d=2)
+      integer x,xq,u,uq
+      integer bxl,bul,bxg,bug
+      integer i
+      integer q
+      parameter(q=6)
 
-      real*8 qref(d*q)
-      real*8 qweight(q)
-      real*8 interp(p*q)
-      real*8 grad(d*p*q)
-      real*8 xq(d*q)
-      real*8 xr(d*p)
-      real*8 iinput(p)
-      real*8 ooutput(q)
-      real*8 val,diff
-      real*8 x1,x2
-      integer*8 ioffset,ooffset
-
-      integer b
+      real*8 p(6)
+      real*8 xx(2)
+      real*8 xxq(q)
+      real*8 uuq(q)
+      real*8 px
+      integer*8 uqoffset,xoffset,offset1,offset2
 
       character arg*32
 
-      xq=(/2.d-1,6.d-1,1.d0/3.d0,2.d-1,2.d-1,2.d-1,   1.d0/3.d0,6.d-1/)
-      xr=(/0.d0,5.d-1,1.d0,0.d0,5.d-1,0.d0,0.d0,0.d0,   0.d0,5.d-1,5.d-1,1.d0/)
+      data p/1,2,3,4,5,6/
+      data xx/-1,1/
 
       call getarg(1,arg)
-
-      call buildmats(qref,qweight,interp,grad)
-
       call ceedinit(trim(arg)//char(0),ceed,err)
 
-      call ceedbasiscreateh1(ceed,ceed_triangle,1,p,q,interp,grad,qref,qweight,&
-     & b,err)
+      call ceedvectorcreate(ceed,2,x,err)
+      xoffset=0
+      call ceedvectorsetarray(x,ceed_mem_host,ceed_use_pointer,xx,xoffset,err)
+      call ceedvectorcreate(ceed,q,xq,err)
+      call ceedvectorsetvalue(xq,0.d0,err)
+      call ceedvectorcreate(ceed,q,u,err)
+      call ceedvectorsetvalue(u,0.d0,err)
+      call ceedvectorcreate(ceed,q,uq,err)
+      call ceedvectorsetvalue(uq,0.d0,err)
 
-      do i=1,p
-        x1=xr(0*p+i)
-        x2=xr(1*p+i)
-        call feval(x1,x2,val)
-        iinput(i)=val
+      call ceedbasiscreatetensorh1lagrange(ceed,1,1,2,q,ceed_gauss_lobatto,&
+     & bxl,err)
+      call ceedbasiscreatetensorh1lagrange(ceed,1,1,q,q,ceed_gauss_lobatto,&
+     & bul,err)
+
+      call ceedbasisapply(bxl,1,ceed_notranspose,ceed_eval_interp,x,xq,err)
+
+      call ceedvectorgetarrayread(xq,ceed_mem_host,xxq,offset1,err)
+      do i=1,q
+        call polyeval(xxq(i+offset1),6,p,uuq(i))
       enddo
-
-      call ceedvectorcreate(ceed,p,input,err)
-      ioffset=0
-      call ceedvectorsetarray(input,ceed_mem_host,ceed_use_pointer,iinput,&
-     & ioffset,err)
-      call ceedvectorcreate(ceed,q,output,err)
-      call ceedvectorsetvalue(output,0.d0,err)
-
-      call ceedbasisapply(b,1,ceed_notranspose,ceed_eval_interp,input,output,&
+      call ceedvectorrestorearrayread(xq,xxq,offset1,err)
+      uqoffset=0
+      call ceedvectorsetarray(uq,ceed_mem_host,ceed_use_pointer,uuq,uqoffset,&
      & err)
 
-      call ceedvectorgetarrayread(output,ceed_mem_host,ooutput,ooffset,err)
+      call ceedbasisapply(bul,1,ceed_transpose,ceed_eval_interp,uq,u,err)
+
+      call ceedbasiscreatetensorh1lagrange(ceed,1,1,2,q,ceed_gauss,bxg,err)
+      call ceedbasiscreatetensorh1lagrange(ceed,1,1,q,q,ceed_gauss,bug,err)
+
+      call ceedbasisapply(bxg,1,ceed_notranspose,ceed_eval_interp,x,xq,err)
+      call ceedbasisapply(bug,1,ceed_notranspose,ceed_eval_interp,u,uq,err)
+
+      call ceedvectorgetarrayread(xq,ceed_mem_host,xxq,offset1,err)
+      call ceedvectorgetarrayread(uq,ceed_mem_host,uuq,offset2,err)
       do i=1,q
-        x1=xq(0*q+i)
-        x2=xq(1*q+i)
-        call feval(x1,x2,val)
-        diff=val-ooutput(i+ooffset)
-        if (abs(diff)>1.0d-10) then
+        call polyeval(xxq(i+offset1),6,p,px)
+        if (abs(uuq(i+offset2)-px) > 1e-14) then
 ! LCOV_EXCL_START
-          write(*,'(A,I1,A,F12.8,A,F12.8)')  '[',i,'] ',ooutput(i+ooffset),&
-     &     ' != ',val
+          write(*,*) uuq(i+offset2),' not eqaul to ',px,'=p(',xxq(i+offset1),')'
 ! LCOV_EXCL_STOP
         endif
       enddo
-      call ceedvectorrestorearrayread(output,ooutput,ooffset,err)
+      call ceedvectorrestorearrayread(xq,xxq,offset1,err)
+      call ceedvectorrestorearrayread(uq,uuq,offest2,err)
 
-      call ceedvectordestroy(input,err)
-      call ceedvectordestroy(output,err)
-      call ceedbasisdestroy(b,err)
+      call ceedvectordestroy(x,err)
+      call ceedvectordestroy(xq,err)
+      call ceedvectordestroy(u,err)
+      call ceedvectordestroy(uq,err)
+      call ceedbasisdestroy(bxl,err)
+      call ceedbasisdestroy(bul,err)
+      call ceedbasisdestroy(bxg,err)
+      call ceedbasisdestroy(bug,err)
       call ceeddestroy(ceed,err)
-
       end
 !-----------------------------------------------------------------------
