+++ tests/t312-basis-f.f90
@@ -1,13 +1,17 @@
 !-----------------------------------------------------------------------
-! 
-! Header with common subroutine
-! 
-      include 't310-basis-f.h'
-!-----------------------------------------------------------------------
-      subroutine feval(x1,x2,val)
-      real*8 x1,x2,val
+      subroutine polyeval(x,n,p,uq)
+      real*8 x,y
+      integer n,i
+      real*8 p(1)
+      real*8 uq
 
-      val=x1*x1+x2*x2+x1*x2+1
+      y=p(n) 
+
+      do i=n-1,1,-1
+        y=y*x+p(i)
+      enddo
+
+      uq=y
 
       end
 !-----------------------------------------------------------------------
@@ -16,82 +20,95 @@
       include 'ceedf.h'
 
       integer ceed,err
-      integer input,output,weights
-      integer p,q,d
-      parameter(p=6)
-      parameter(q=4)
-      parameter(d=2)
+      integer x,xq,u,uq,w
+      integer bxl,bxg,bug
+      integer i
+      integer q
+      parameter(q=6)
 
-      real*8 qref(d*q)
-      real*8 qweight(q)
-      real*8 interp(p*q)
-      real*8 grad(d*p*q)
-      real*8 xr(d*p)
-      real*8 iinput(p)
-      real*8 ooutput(q)
-      real*8 wweights(q)
-      real*8 val,diff
-      real*8 x1,x2
-      integer*8 ioffset,offset1,offset2
-
-      integer b
+      integer plen
+      parameter(plen=6)
+      real*8 p(plen)
+      real*8 pint(plen+1)
+      real*8 xx(2)
+      real*8 xxq(q)
+      real*8 uuq(q)
+      real*8 uu(q)
+      real*8 ww(q)
+      real*8 summ,error,p1,pm1
+      integer*8 uoffset,xoffset,offset1,offset2
 
       character arg*32
 
-      xr=(/0.0d0,5.0d-1,1.0d0,0.0d0,5.0d-1,0.0d0,0.0d0,0.0d0,0.0d0,5.0d-1,&
-     &  5.0d-1,1.0d0/)
+      data p/1,2,3,4,5,6/
+      data xx/-1,1/
 
       call getarg(1,arg)
-
-      call buildmats(qref,qweight,interp,grad)
-
       call ceedinit(trim(arg)//char(0),ceed,err)
 
-      call ceedbasiscreateh1(ceed,ceed_triangle,1,p,q,interp,grad,qref,qweight,&
-     & b,err)
+      call ceedvectorcreate(ceed,2,x,err)
+      xoffset=0
+      call ceedvectorsetarray(x,ceed_mem_host,ceed_use_pointer,xx,xoffset,err)
+      call ceedvectorcreate(ceed,q,xq,err)
+      call ceedvectorsetvalue(xq,0.d0,err)
+      call ceedvectorcreate(ceed,q,u,err)
+      call ceedvectorcreate(ceed,q,uq,err)
+      call ceedvectorsetvalue(uq,0.d0,err)
+      call ceedvectorcreate(ceed,q,w,err)
+      call ceedvectorsetvalue(w,0.d0,err)
 
-      do i=1,p
-        x1=xr(0*p+i)
-        x2=xr(1*p+i)
-        call feval(x1,x2,val)
-        iinput(i)=val
-      enddo
-
-      call ceedvectorcreate(ceed,p,input,err)
-      ioffset=0
-      call ceedvectorsetarray(input,ceed_mem_host,ceed_use_pointer,iinput,&
-     & ioffset,err)
-      call ceedvectorcreate(ceed,q,output,err)
-      call ceedvectorsetvalue(output,0.d0,err)
-      call ceedvectorcreate(ceed,q,weights,err)
-      call ceedvectorsetvalue(weights,0.d0,err)
-
-      call ceedbasisapply(b,1,ceed_notranspose,ceed_eval_interp,input,output,&
+      call ceedbasiscreatetensorh1lagrange(ceed,1,1,2,q,ceed_gauss_lobatto,bxl,&
      & err)
-      call ceedbasisapply(b,1,ceed_notranspose,ceed_eval_weight,ceed_null,&
-     & weights,err)
 
-      call ceedvectorgetarrayread(output,ceed_mem_host,ooutput,offset1,err)
-      call ceedvectorgetarrayread(weights,ceed_mem_host,wweights,offset2,err)
-      val=0
+      call ceedbasisapply(bxl,1,ceed_notranspose,ceed_eval_interp,x,xq,err)
+
+      call ceedvectorgetarrayread(xq,ceed_mem_host,xxq,offset1,err)
       do i=1,q
-        val=val+ooutput(i+offset1)*wweights(i+offset2)
+        call polyeval(xxq(i+offset1),plen,p,uu(i))
       enddo
-      call ceedvectorrestorearrayread(output,ooutput,offset1,err)
-      call ceedvectorrestorearrayread(weights,wweights,offset2,err)
+      call ceedvectorrestorearrayread(xq,xxq,offset1,err)
+      uoffset=0
+      call ceedvectorsetarray(u,ceed_mem_host,ceed_use_pointer,uu,uoffset,err)
 
-      diff=val-17.d0/24.d0
-      if (abs(diff)>1.0d-10) then
+      call ceedbasiscreatetensorh1lagrange(ceed,1,1,2,q,ceed_gauss,bxg,err)
+      call ceedbasiscreatetensorh1lagrange(ceed,1,1,q,q,ceed_gauss,bug,err)
+
+      call ceedbasisapply(bxg,1,ceed_notranspose,ceed_eval_interp,x,xq,err)
+      call ceedbasisapply(bug,1,ceed_notranspose,ceed_eval_interp,u,uq,err)
+      call ceedbasisapply(bug,1,ceed_notranspose,ceed_eval_weight,ceed_null,w,&
+     & err)
+
+      call ceedvectorgetarrayread(w,ceed_mem_host,ww,offset1,err)
+      call ceedvectorgetarrayread(uq,ceed_mem_host,uuq,offset2,err)
+      summ=0.0
+      do i=1,q
+        summ=summ+ww(i+offset1)*uuq(i+offset2)
+      enddo
+      call ceedvectorrestorearrayread(w,ww,offset1,err)
+      call ceedvectorrestorearrayread(uq,uuq,offset2,err)
+
+      pint(1)=0.0
+      do i=1,plen
+        pint(i+1)=p(i)/i
+      enddo
+
+      call polyeval(1.0D0,plen,pint,p1)
+      call polyeval(-1.0D0,plen,pint,pm1)
+      error=summ-p1+pm1
+      if (abs(error) > 1e-10) then
 ! LCOV_EXCL_START
-        write(*,'(A,I1,A,F12.8,A,F12.8)')'[',i,'] ',val,' != ',17.d0/24.d0
+        write(*,*) 'Error ',error,' sum ',summ,' exact ',p1-pm1
 ! LCOV_EXCL_STOP
       endif
 
-      call ceedvectordestroy(input,err)
-      call ceedvectordestroy(output,err)
-      call ceedvectordestroy(weights,err)
-      call ceedbasisdestroy(b,err)
+      call ceedvectordestroy(x,err)
+      call ceedvectordestroy(xq,err)
+      call ceedvectordestroy(u,err)
+      call ceedvectordestroy(uq,err)
+      call ceedvectordestroy(w,err)
+      call ceedbasisdestroy(bxl,err)
+      call ceedbasisdestroy(bxg,err)
+      call ceedbasisdestroy(bug,err)
       call ceeddestroy(ceed,err)
-
       end
 !-----------------------------------------------------------------------
