+++ tests/t312-basis.c
@@ -1,61 +1,82 @@
 /// @file
-/// Test integration with a 2D Simplex non-tensor H1 basis
-/// \test Test integration with a 2D Simplex non-tensor H1 basis
+/// Test polynomial interpolation in 1D
+/// \test Test polynomial interpolation in 1D
 #include <ceed.h>
 #include <math.h>
-#include "t310-basis.h"
 
-double feval(double x1, double x2) {
-  return x1*x1 + x2*x2 + x1*x2 + 1;
+#define ALEN(a) (sizeof(a) / sizeof((a)[0]))
+
+static CeedScalar PolyEval(CeedScalar x, CeedInt n, const CeedScalar *p) {
+  CeedScalar y = p[n-1];
+  for (CeedInt i=n-2; i>=0; i--) y = y*x + p[i];
+  return y;
 }
 
 int main(int argc, char **argv) {
   Ceed ceed;
-  CeedVector In, Out, Weights;
-  const CeedInt P = 6, Q = 4, dim = 2;
-  CeedBasis b;
-  CeedScalar qref[dim*Q], qweight[Q];
-  CeedScalar interp[P*Q], grad[dim*P*Q];
-  CeedScalar xr[] = {0., 0.5, 1., 0., 0.5, 0., 0., 0., 0., 0.5, 0.5, 1.};
-  const CeedScalar *out, *weights;
-  CeedScalar in[P], sum;
-
-  buildmats(qref, qweight, interp, grad);
+  CeedVector X, Xq, U, Uq, W;
+  CeedBasis bxl, bxg, bug;
+  CeedInt Q = 6;
+  const CeedScalar p[6] = {1, 2, 3, 4, 5, 6}; // 1 + 2x + 3x^2 + ...
+  const CeedScalar *xq, *uq, *w;
+  CeedScalar u[Q], x[2], sum, error, pint[ALEN(p)+1];
 
   CeedInit(argv[1], &ceed);
 
-  CeedBasisCreateH1(ceed, CEED_TRIANGLE, 1, P, Q, interp, grad, qref, qweight,
-                    &b);
+  CeedVectorCreate(ceed, 2, &X);
+  CeedVectorCreate(ceed, Q, &Xq);
+  CeedVectorSetValue(Xq, 0);
+  CeedVectorCreate(ceed, Q, &U);
+  CeedVectorCreate(ceed, Q, &Uq);
+  CeedVectorSetValue(Uq, 0);
+  CeedVectorCreate(ceed, Q, &W);
+  CeedVectorSetValue(W, 0);
 
-  // Interpolate function to quadrature points
-  for (int i=0; i<P; i++)
-    in[i] = feval(xr[0*P+i], xr[1*P+i]);
+  CeedBasisCreateTensorH1Lagrange(ceed, 1,  1, 2, Q, CEED_GAUSS_LOBATTO, &bxl);
 
-  CeedVectorCreate(ceed, P, &In);
-  CeedVectorSetArray(In, CEED_MEM_HOST, CEED_USE_POINTER, (CeedScalar *)&in);
-  CeedVectorCreate(ceed, Q, &Out);
-  CeedVectorSetValue(Out, 0);
-  CeedVectorCreate(ceed, Q, &Weights);
-  CeedVectorSetValue(Weights, 0);
+  for (int i = 0; i < 2; i++) x[i] = CeedIntPow(-1, i+1);
+  CeedVectorSetArray(X, CEED_MEM_HOST, CEED_USE_POINTER, (CeedScalar *)&x);
 
-  CeedBasisApply(b, 1, CEED_NOTRANSPOSE, CEED_EVAL_INTERP, In, Out);
-  CeedBasisApply(b, 1, CEED_NOTRANSPOSE, CEED_EVAL_WEIGHT, NULL, Weights);
+  CeedBasisApply(bxl, 1, CEED_NOTRANSPOSE, CEED_EVAL_INTERP, X, Xq);
 
-  // Check values at quadrature points
-  CeedVectorGetArrayRead(Out, CEED_MEM_HOST, &out);
-  CeedVectorGetArrayRead(Weights, CEED_MEM_HOST, &weights);
+  CeedVectorGetArrayRead(Xq, CEED_MEM_HOST, &xq);
+  for (CeedInt i=0; i<Q; i++) u[i] = PolyEval(xq[i], ALEN(p), p);
+  CeedVectorRestoreArrayRead(Xq, &xq);
+  CeedVectorSetArray(U, CEED_MEM_HOST, CEED_USE_POINTER, (CeedScalar *)&u);
+
+  CeedBasisCreateTensorH1Lagrange(ceed, 1, 1, 2, Q, CEED_GAUSS, &bxg);
+  CeedBasisCreateTensorH1Lagrange(ceed, 1, 1, Q, Q, CEED_GAUSS, &bug);
+
+  CeedBasisApply(bxg, 1, CEED_NOTRANSPOSE, CEED_EVAL_INTERP, X, Xq);
+  CeedBasisApply(bug, 1, CEED_NOTRANSPOSE, CEED_EVAL_INTERP, U, Uq);
+  CeedBasisApply(bug, 1, CEED_NOTRANSPOSE, CEED_EVAL_WEIGHT, NULL, W);
+
+  CeedVectorGetArrayRead(W, CEED_MEM_HOST, &w);
+  CeedVectorGetArrayRead(Uq, CEED_MEM_HOST, &uq);
   sum = 0;
-  for (int i=0; i<Q; i++)
-    sum += out[i]*weights[i];
-  if (fabs(sum - 17./24.) > 1e-10)
-    printf("%f != %f\n", sum, 17./24.);
-  CeedVectorRestoreArrayRead(Out, &out);
-  CeedVectorRestoreArrayRead(Weights, &weights);
+  for (CeedInt i=0; i<Q; i++) {
+    sum += w[i] * uq[i];
+  }
+  CeedVectorRestoreArrayRead(W, &w);
+  CeedVectorRestoreArrayRead(Uq, &uq);
 
-  CeedVectorDestroy(&In);
-  CeedVectorDestroy(&Out);
-  CeedVectorDestroy(&Weights);
-  CeedBasisDestroy(&b);
+  pint[0] = 0;
+  for (CeedInt i=0; i<(int)ALEN(p); i++) pint[i+1] = p[i] / (i+1);
+  error = sum - PolyEval(1, ALEN(pint), pint) + PolyEval(-1, ALEN(pint), pint);
+  if (error > 1.e-10)
+    // LCOV_EXCL_START
+    printf("Error %e  sum %g  exact %g\n", error, sum,
+           PolyEval(1, ALEN(pint), pint) - PolyEval(-1, ALEN(pint), pint));
+    // LCOV_EXCL_STOP
+
+  CeedVectorDestroy(&X);
+  CeedVectorDestroy(&Xq);
+  CeedVectorDestroy(&U);
+  CeedVectorDestroy(&Uq);
+  CeedVectorDestroy(&W);
+  CeedBasisDestroy(&bxl);
+  CeedBasisDestroy(&bxg);
+  CeedBasisDestroy(&bug);
   CeedDestroy(&ceed);
   return 0;
 }
