+++ tests/t316-basis-f.f90
@@ -26,7 +26,7 @@
       integer p
       integer q
       parameter(p=8)
-      parameter(q=10)
+      parameter(q=7)
       integer maxdim
       parameter(maxdim=3)
       integer qdimnmax
