+++ tests/t511-operator-f.f90
@@ -2,7 +2,7 @@
 ! 
 ! Header with common subroutine
 ! 
-      include 't310-basis-f.h'
+      include 't320-basis-f.h'
 !-----------------------------------------------------------------------
       subroutine setup(ctx,q,u1,u2,u3,u4,u5,u6,u7,u8,u9,u10,u11,u12,u13,u14,&
 &           u15,u16,v1,v2,v3,v4,v5,v6,v7,v8,v9,v10,v11,v12,v13,v14,v15,v16,ierr)
