+++ tests/t521-operator.c
@@ -4,7 +4,7 @@
 #include <ceed.h>
 #include <stdlib.h>
 #include <math.h>
-#include "t310-basis.h"
+#include "t320-basis.h"
 
 #include "t521-operator.h"
