+++ tests/tap.sh
@@ -71,9 +71,9 @@ for ((i=0;i<${#backends[@]}; ++i)); do
         continue
     fi
 
-    # grep to pass test t308 on error
+    # grep to pass test t303 on error
     if grep -F -q -e 'vectors incompatible' ${output}.err \
-            && [[ "$1" = "t308"* ]] ; then
+            && [[ "$1" = "t303"* ]] ; then
         printf "ok $i0 PASS - expected failure $1 $backend\n"
         printf "ok $i1 PASS - expected failure $1 $backend stdout\n"
         printf "ok $i2 PASS - expected failure $1 $backend stderr\n"
